Diagnostics cmdlet update
Changes:
• IaaS and PaaS Set cmdlet now take a full XML config as described in
the schema.
• PaaS Set and New cmdlet DiagnosticsConfiguration parameter has been
renamed DiagnosticsConfigurationPath and the type is now string instead
of XmlDocument. This is to be consistent with the IaaS parameter of the
same name (this is the config).
• PaaS cmdlet Set and New StorageAccountName parameter has been replaced
by StorageContext of type AzureStorageContext. This is to be consistent
with IaaS and it is the new way of passing account credentials.
• IaaS Remove now throws a NotSupportedException as it is not fully
implemented on the rdfe side.